+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.3.0] - 2025-02-12
+
+### Added 
+- Capability for user to specify fields of a `LoadableConfig` as required.
+
+### Fixed
+- Bug in `TunedJoystick` where `setPeriodMilliseconds()` was not declared public.
+
 ## [0.2.0] - 2025-01-11
 
 ### Fixed